requirements.txt
torch
diffusers
Pillow
streamlit
transformers